Force number of bins to be an integer.

+++ seaborn/distributions.py
@@ -27,9 +27,9 @@ def _freedman_diaconis_bins(a):
     h = 2 * iqr(a) / (len(a) ** (1 / 3))
     # fall back to sqrt(a) bins if iqr is 0
     if h == 0:
-        return np.sqrt(a.size)
+        return int(np.sqrt(a.size))
     else:
-        return np.ceil((a.max() - a.min()) / h)
+        return int(np.ceil((a.max() - a.min()) / h))
 
 
 def distplot(a, bins=None, hist=True, kde=True, rug=False, fit=None,
